requirements_full_ml.txt
# Полные зависимости для максимальной точности AI модели
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.8.0
requests>=2.31.0
langdetect>=1.0.9
python-multipart>=0.0.6
aiogram>=3.0.0
aiohttp>=3.9.0

# ML библиотеки для максимальной точности
numpy>=1.24.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.1
nltk>=3.8.1
pandas>=2.0.0
scikit-learn>=1.3.0

# Дополнительные библиотеки для обработки текста
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
